# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  garden.CreateAccessLinkRequest:
    properties:
      expires_at:
        type: string
      garden_id:
        type: string
      max_uses:
        type: integer
      permissions:
        items:
          $ref: '#/definitions/garden.GardenPermission'
        type: array
    required:
    - garden_id
    - permissions
    type: object
...
  garden.GardenPermission:
    enum:
    - view
    - plant
    - harvest
    - manage
    type: string
    x-enum-varnames:
    - GardenPermissionView
    - GardenPermissionPlant
    - GardenPermissionHarvest
    - GardenPermissionManage
  garden.JoinGardenRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
...
  garden.UpdateSharePermissionsRequest:
    properties:
      permissions:
        items:
          $ref: '#/definitions/garden.GardenPermission'
        type: array
      user_id:
        type: string
    required:
    - permissions
    - user_id
    type: object
...
info:
  contact: {}
  description: A robust RESTful API for a virtual garden management game built with
    Go.
  title: Virtual Garden Management System API
  version: "1.0"
paths:
...
  /garden-shares/access-links:
    post:
      consumes:
      - application/json
      description: Create a shareable link for garden access with specified permissions
      parameters:
      - description: Access link creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/garden.CreateAccessLinkRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created access link
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Not garden owner
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Garden not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Create garden access link
      tags:
      - garden-sharing
  /garden-shares/garden/{garden_id}:
    get:
      consumes:
      - application/json
      description: Get all users who have access to a specific garden
      parameters:
      - description: Garden ID
        in: path
        name: garden_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of garden shares
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Not garden owner
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Garden not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Get garden shares
      tags:
      - garden-sharing
  /garden-shares/garden/{garden_id}/access-links:
    get:
      consumes:
      - application/json
      description: Get all access links for a garden
      parameters:
      - description: Garden ID
        in: path
        name: garden_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of access links
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Not garden owner
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Garden not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Get access links
      tags:
      - garden-sharing
  /garden-shares/garden/{garden_id}/access-links/{link_id}/deactivate:
    post:
      consumes:
      - application/json
      description: Deactivate an access link for a garden
      parameters:
      - description: Garden ID
        in: path
        name: garden_id
        required: true
        type: string
      - description: Access Link ID
        in: path
        name: link_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Access link deactivated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Not garden owner
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Access link not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Deactivate access link
      tags:
      - garden-sharing
  /garden-shares/garden/{garden_id}/permissions:
    put:
      consumes:
      - application/json
      description: Update permissions for a user's access to a garden
      parameters:
      - description: Garden ID
        in: path
        name: garden_id
        required: true
        type: string
      - description: Permission update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/garden.UpdateSharePermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated garden share
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Not garden owner
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Garden share not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Update share permissions
      tags:
      - garden-sharing
  /garden-shares/garden/{garden_id}/user/{user_id}:
    delete:
      consumes:
      - application/json
      description: Remove a user's access to a garden
      parameters:
      - description: Garden ID
        in: path
        name: garden_id
        required: true
        type: string
      - description: User ID to remove
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Garden share removed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Not garden owner
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Garden share not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Remove garden share
      tags:
      - garden-sharing
  /garden-shares/join:
    post:
      consumes:
      - application/json
      description: Join a garden using an access link token
      parameters:
      - description: Join garden data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/garden.JoinGardenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully joined garden
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Access link not found or expired
          schema:
            additionalProperties: true
            type: object
        "410":
          description: Access link expired or max uses reached
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Join garden via access link
      tags:
      - garden-sharing
  /garden-shares/shared-with-me:
    get:
      consumes:
      - application/json
      description: Get all gardens shared with the current user
      produces:
      - application/json
      responses:
        "200":
          description: List of shared gardens
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Get shared gardens
      tags:
      - garden-sharing
...
securityDefinitions:
  bearer:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"